Add files via upload
This code includes the SheetJS and FileSaver libraries in the <head> section of the HTML file using the <script> tag. The <input> and <button> elements are added to the <body> section of the HTML file. The convert() and s2ab() functions are defined in a <script> tag in the <body> section of the HTML file.

Note that this code uses the saveAs function from the FileSaver.js library (https://github.com/eligrey/FileSaver.js/) to download the converted file. You'll need to include this library in your HTML page to use the saveAs function.